requirements.txt
ultralytics>=8.3.156
pyyaml